+++ docs/examples/e06_simple_bot_structure/main.py
@@ -16,7 +16,7 @@ async def ready(client):
 
 # Registered variables to extension loader, show up in each loaded extension file.
 EXTENSION_LOADER.add_default_variables(Sakuya=Sakuya)
-# Adds the extensions file or the extensions files recursive in the folder
+# Adds the extensions file or the extensions files recursive in the directory
 EXTENSION_LOADER.add('modules')
 # Loads all the added extension files.
 EXTENSION_LOADER.load_all()