+++ hata/ext/extension_loader/extension_loader.py
@@ -1,7 +1,6 @@
 __all__ = ('EXTENSION_LOADER', 'ExtensionLoader', )
 
 from functools import partial as partial_func
-from importlib import __file__ as IMPORTLIB_FILE_PATH
 
 from scarletio import (
     HybridValueDictionary, RichAttributeErrorBaseType, Task, alchemy_incendiary, export, is_coroutine_function,
@@ -13,6 +12,7 @@ from ...discord.core import KOKORO
 from .constants import EXTENSION_STATE_LOADED, EXTENSIONS
 from .exceptions import ExtensionError
 from .extension import Extension, __file__ as EXTENSION_LOADER_EXTENSION_FILE_PATH
+from .extension_root import maybe_register_extension_directory_root
 from .helpers import (
     PROTECTED_NAMES, _build_extension_tree, _get_extension_name_and_path, _get_path_extension_name,
     _iter_extension_names_and_paths, _validate_entry_or_exit, validate_extension_parameters
@@ -173,41 +173,37 @@ def _ignore_module_import_frames(file_name, name, line_number, line):
     should_show_frame : `bool`
         Whether the frame should be shown.
     """
+    should_show_frame = True
+    
     if file_name.startswith('<') and file_name.endswith('>'):
-        return False
+        should_show_frame = False
     
-    if file_name == EXTENSION_LOADER_EXTENSION_FILE_PATH:
+    elif file_name == EXTENSION_LOADER_EXTENSION_FILE_PATH:
         if name == '_load':
-            if line in (
-                'reload_module(module)',
-                'spec.loader.exec_module(module)',
-            ):
-                return False
+            if line == 'loaded = self._load_module()':
+                should_show_frame = False
+        
+        elif name == '_load_module':
+            if line == 'spec.loader.exec_module(module)':
+                should_show_frame = False
     
-    if file_name == EXTENSION_LOADER_EXTENSION_LOADER_FILE_PATH:
+    elif file_name == EXTENSION_LOADER_EXTENSION_LOADER_FILE_PATH:
         if name == '_extension_loader_task':
             if line in (
                 'module = await KOKORO.run_in_executor(extension._load)',
                 'await entry_point(module)',
                 'entry_point(module)',
             ):
-                return False
+                should_show_frame = False
         
-        if name == '_extension_unloader_task':
+        elif name == '_extension_unloader_task':
             if line in (
                 'await exit_point(module)',
                 'exit_point(module)',
             ):
-                return False
-            
-            return False
+                should_show_frame = False
     
-    # Importlib paths might contain important information about why python derps out when we trying to do something
-    # smarter as they ever expected!
-    # if file_name == IMPORTLIB_FILE_PATH:
-    #     return False
-    
-    return True
+    return should_show_frame
 
 
 def _pop_unloader_task_callback(extension, task):
@@ -742,13 +738,17 @@ class ExtensionLoader(RichAttributeErrorBaseType):
         ValueError
             If a variable name is would be used, what is `module` attribute.
         """
-        extension_names_and_paths = set(_iter_extension_names_and_paths(name))
+        extension_names_and_paths = set(_iter_extension_names_and_paths(name, register_directories_as_roots=True))
         entry_point, exit_point, extend_default_variables, locked, take_snapshot_difference, default_variables = \
             validate_extension_parameters(*args, **kwargs)
         
+        maybe_register_extension_directory_root(name)
+        
         for extension_name, extension_path in extension_names_and_paths:
-            self._add(extension_name, extension_path, entry_point, exit_point, extend_default_variables, locked,
-                take_snapshot_difference, default_variables)
+            self._add(
+                extension_name, extension_path, entry_point, exit_point, extend_default_variables, locked,
+                take_snapshot_difference, default_variables
+            )
     
     
     def _add(self, extension_name, extension_path, entry_point, exit_point, extend_default_variables, locked,