+++ hata/ext/extension_loader/helpers.py
@@ -2,13 +2,15 @@ __all__ = ()
 
 from os import listdir as list_directory
 from os.path import (
-    basename as base_name, exists, isabs as is_absolute_path_name, isdir as is_folder, isfile as is_file,
+    basename as base_name, exists, isabs as is_absolute_path_name, isdir as is_directory, isfile as is_file,
     join as join_paths
 )
 from sys import path as route_paths
 
 from scarletio import CallableAnalyzer, HybridValueDictionary
 
+from .constants import ABSOLUTE_PATH_EXTENSION_NAME_PREFIX
+from .extension_root import register_extension_root
 
 def _validate_entry_or_exit(point):
     """
@@ -69,7 +71,7 @@ def validate_extension_parameters(entry_point=None, exit_point=None, extend_defa
     exit_point : `None`, `str`, `callable` = `None`, Optional
         Extension specific exit point, to use over the extension loader's default.
     extend_default_variables : `bool` = `True`, Optional
-        Whether the extension should use the loader's default variables or just it's own's.
+        Whether the extension should use the loader's default variables or just it's own.
     locked : `bool` = `False`, Optional
         Whether the given extension(s) should not be affected by `.{}_all` methods.
     take_snapshot_difference : `bool` = `True`, Optional
@@ -99,7 +101,7 @@ def validate_extension_parameters(entry_point=None, exit_point=None, extend_defa
     exit_point : `None`, `str`, `callable`
         Extension specific exit point, to use over the extension loader's default.
     extend_default_variables : `bool`
-        Whether the extension should use the loader's default variables or just it's own's.
+        Whether the extension should use the loader's default variables or just it's own.
     locked : `bool`
         Whether the given extension(s) should not be affected by `.{}_all` methods.
     take_snapshot_difference : `bool`
@@ -155,55 +157,17 @@ def validate_extension_parameters(entry_point=None, exit_point=None, extend_defa
     return entry_point, exit_point, extend_default_variables, locked, take_snapshot_difference, default_variables
 
 
-PROTECTED_NAMES = {
-    '__class__',
-    '__delattr__',
-    '__dict__',
-    '__dir__',
-    '__doc__',
-    '__eq__',
-    '__format__',
-    '__ge__',
-    '__getattribute__',
-    '__gt__',
-    '__hash__',
-    '__init__',
-    '__init_subclass__',
-    '__le__',
-    '__lt__',
-    '__module__',
-    '__ne__',
-    '__new__',
-    '__reduce__',
-    '__reduce_ex__',
-    '__repr__',
-    '__setattr__',
-    '__sizeof__',
-    '__str__',
-    '__subclasshook__',
-    '__weakref__',
-    '_cached',
-    '_set_fileattr',
-    'cached',
-    'has_location',
-    'loader',
-    'loader_state',
-    'name',
-    'origin',
-    'parent',
-    'submodule_search_locations',
-    '__path__',
-    '__spec__',
-}
+PROTECTED_NAMES = frozenset((
+    '__class__', '__delattr__', '__dict__', '__dir__', '__doc__', '__eq__', '__format__', '__ge__', '__getattribute__',
+    '__gt__', '__hash__', '__init__', '__init_subclass__', '__le__', '__lt__', '__module__', '__ne__', '__new__',
+    '__reduce__', '__reduce_ex__', '__repr__', '__setattr__', '__sizeof__', '__str__', '__subclasshook__',
+    '__weakref__', '_cached', '_set_fileattr', 'cached', 'has_location', 'loader', 'loader_state', 'name', 'origin',
+    'parent', 'submodule_search_locations', '__path__', '__spec__'
+))
 
 
 
-PYTHON_EXTENSION_NAMES = (
-    '.py',
-    '.pyd',
-    '.pyc',
-    '.so',
-)
+PYTHON_EXTENSION_NAMES = frozenset(('.py', '.pyd', '.pyc', '.so'))
 
 
 def _get_extension_name_and_path(name):
@@ -250,7 +214,7 @@ def _get_extension_name_and_path(name):
     return extension_pair
 
 
-def _iter_extension_names_and_paths(name):
+def _iter_extension_names_and_paths(name, *, register_directories_as_roots=False):
     """
     Fetches the names and the paths of the given extension.
     
@@ -260,6 +224,8 @@ def _iter_extension_names_and_paths(name):
     ----------
     name : `str`, `iterable` of `str`
         The name(s) to fetch.
+    register_directories_as_roots : `bool` = `False`, Optional (Keyword only)
+        Whether directory roots should be registered.
     
     Yields
     ------
@@ -282,7 +248,7 @@ def _iter_extension_names_and_paths(name):
             yield name
             return
         
-        yield from _lookup_path(name)
+        yield from _lookup_path(name, register_directories_as_roots)
 
 
 def _iter_name_maybe_iterable(name):
@@ -313,26 +279,31 @@ def _iter_name_maybe_iterable(name):
     name_type = type(name)
     if name_type is str:
         yield name
+    
     elif issubclass(name_type, str):
         yield str(name)
+    
     elif hasattr(name_type, '__iter__'):
         for sub_name in name:
             sub_name_type = type(sub_name)
             if sub_name_type is str:
                 yield sub_name
+            
             elif issubclass(sub_name_type, str):
                 yield str(sub_name)
+            
             else:
                 raise TypeError(
                     f'`name` contains a non `str` element, got {sub_name_type.__name__}; {sub_name!r}; name={name!r}.'
                 )
+    
     else:
         raise TypeError(
             f'`name` can be `str`, `iterable` of `str`, got {name_type.__name__}; {name!r}.'
         )
 
 
-def _lookup_path(import_name_or_path):
+def _lookup_path(import_name_or_path, register_directories_as_roots):
     """
     Detects the root of the given name.
     
@@ -342,6 +313,8 @@ def _lookup_path(import_name_or_path):
     ----------
     import_name_or_path : `str`
         An extension's import name, or it's absolute path.
+    register_directories_as_roots : `bool`
+        Whether directory roots should be registered.
     
     Yields
     ------
@@ -357,8 +330,8 @@ def _lookup_path(import_name_or_path):
     """
     if is_absolute_path_name(import_name_or_path):
         if exists(import_name_or_path):
-            if is_folder(import_name_or_path):
-                yield from _iter_folder(None, import_name_or_path)
+            if is_directory(import_name_or_path):
+                yield from _iter_directory(None, import_name_or_path)
                 return
             
             if is_file(import_name_or_path):
@@ -368,8 +341,10 @@ def _lookup_path(import_name_or_path):
         path_end = join_paths(*import_name_or_path.split('.'))
         for base_path in route_paths:
             path = join_paths(base_path, path_end)
-            if exists(path) and is_folder(path):
-                yield from _iter_folder(import_name_or_path, path)
+            if exists(path) and is_directory(path):
+                if register_directories_as_roots:
+                    register_extension_root(import_name_or_path)
+                yield from _iter_directory(import_name_or_path, path)
                 return
             
             for python_extension_name in PYTHON_EXTENSION_NAMES:
@@ -384,35 +359,35 @@ def _lookup_path(import_name_or_path):
     )
 
 
-def _iter_folder(import_name, folder_path):
+def _iter_directory(import_name, directory_path):
     """
-    Iterates over a folder's import names.
+    Iterates over a directory's import names.
     
     Parameters
     ----------
     extension_name : `None`, `str`
         The name of the extension if we would import it.
-    folder_path : `str`
-        Path to the folder
+    directory_path : `str`
+        Path to the directory
     
     Yields
     ------
     extension_name : `None`, `str`
-        Detected import names for each applicable file in the folder.
+        Detected import names for each applicable file in the directory.
     extension_path : `str`
         Path of the file.
     """
     for python_extension_name in PYTHON_EXTENSION_NAMES:
-        file_path = join_paths(folder_path, f'__init__{python_extension_name}')
+        file_path = join_paths(directory_path, f'__init__{python_extension_name}')
         if exists(file_path) and is_file(file_path):
             yield import_name, file_path
             return
     
-    for file_name in list_directory(folder_path):
+    for file_name in list_directory(directory_path):
         if file_name.startswith('.') or (file_name == '__pycache__'):
             continue
         
-        path = join_paths(folder_path, file_name)
+        path = join_paths(directory_path, file_name)
         
         if is_file(path):
             for python_extension_name in PYTHON_EXTENSION_NAMES:
@@ -426,18 +401,17 @@ def _iter_folder(import_name, folder_path):
             
             continue
         
-        if is_folder(path):
+        if is_directory(path):
             if import_name is None:
                 import_name_value = None
             else:
                 import_name_value = f'{import_name}.{file_name}'
-            yield from _iter_folder(import_name_value, path)
+            yield from _iter_directory(import_name_value, path)
             continue
         
         # no more cases
         continue
 
-ABSOLUTE_PATH_EXTENSION_NAME_PREFIX = '<extension>.'
 
 def _get_path_extension_name(path):
     """